# swagger/docs/main.yml
openapi: 3.0.0
...
info:
  version: "1.0.0"
  title: Chortech Api Documentation
  description: >-
    Chortech is an splitwise clone application
...
paths:
...
  /api/expenses/{id}/comments:
    post:
      security:
        - bearerAuth: []
      tags:
        - "expense"
      summary: Add a comment
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      description: >-
        Add a comment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                created_at:
                  type: number
                  example: 1608843422
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment added."
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              examples:
                validation:
                  value:
                    errors:
                      - message: "text must be less than 255 characters"
                        fields: ["text"]
                no-part:
                  value:
                    errors:
                      - message: "user doesn't participate in expense"
                unknown:
                  value:
                    errors:
                      - message: "Something went wrong"
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "Expenese doesn't exists!"
    get:
      security:
        - bearerAuth: []
      tags:
        - "expense"
      summary: Get all comments
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
          required: true
      description: >-
        Get all comments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    text:
                      type: string
                    created_at:
                      type: number
                    id:
                      type: string
                      format: uuid
                    writer:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
        "401":
          description: UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "Invalid authentication credentials provided!"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "You are not authorized to access this resource!"
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "Expenese doesn't exists!"
  /api/payments:
    post:
      security:
        - bearerAuth: []
      tags:
        - "payment"
      summary: Create Payment
      description: >-
        Create a payment between two uses
      requestBody:
        description: this is the request body
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - to
                - amount
              properties:
                from:
                  type: string
                to:
                  type: string
                amount:
                  type: number
                paid_at:
                  type: number
                group:
                  type: string
                notes:
                  type: string
      responses:
        "201":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              examples:
                validation:
                  value:
                    errors:
                      - message: "from is required!"
                        fields: ["from"]
                unknown:
                  value:
                    errors:
                      - message: "Something went wrong"
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "Group does not exists!"
  /api/payments/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - "payment"
      summary: Payment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      description: >-
        Get a payment
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "Didn't found payment with given id!"
    delete:
      security:
        - bearerAuth: []
      tags:
        - "payment"
      summary: Delete Payment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      description: >-
        Delete a payment between two uses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Delete successful"
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "Group does not exists!"
    put:
      security:
        - bearerAuth: []
      tags:
        - "payment"
      summary: Update Payment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      description: >-
        Update a payment between two uses note that you can't change neither from nor to
      requestBody:
        description: this is the request body
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                paid_at:
                  type: number
                group:
                  type: string
                notes:
                  type: string
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              examples:
                validation:
                  value:
                    errors:
                      - message: "darivari is not allowed!"
                        fields: ["darivari"]
                unknown:
                  value:
                    errors:
                      - message: "Something went wrong"
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "user doesn't participate in payment"
  /api/payments/{id}/comments:
    post:
      security:
        - bearerAuth: []
      tags:
        - "payment"
      summary: Add comment on Payment
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      description: >-
        Add a comment on Payment
      requestBody:
        description: this is the request body
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - created_at
              properties:
                text:
                  type: string
                created_at:
                  type: string
      responses:
        "201":
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment added."
        "404":
          description: NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
              example:
                errors:
                  - message: "User doesn't participate in payment"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      properties:
        id:
          type: string
        from:
          $ref: "#/components/schemas/User"
        to:
          $ref: "#/components/schemas/User"
        group:
          $ref: "#/components/schemas/Group"
        amount:
          type: number
        paid_at:
          type: string
        notes:
          type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
...
    Comment:
      properties:
        id:
          type: string
        text:
          type: string
        created_at:
          type: number
        writer:
          $ref: "#/components/schemas/User"
    Group:
      properties:
        id:
          type: string
        name:
          type: string
        picture:
          type: string
        creator:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
    User:
      properties:
        id:
          type: string
        picture:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    Errors:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
              fields:
                nullable: true
                type: array
                items:
                  type: string